requirements.txt
# Azure Functions runtime
azure-functions

# CosmosDB SDK
azure-cosmos

# Additional utilities
python-dotenv
pydantic

# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry